# Fixes slow query in notifications api